Properties cleanup - part 1
