+++ bootstrap/frontend/src/main/scala/io/udash/bootstrap/button/UdashButtonToolbar.scala
@@ -3,13 +3,14 @@ package button
 
 import io.udash._
 import io.udash.bootstrap.UdashBootstrap.ComponentId
+import io.udash.properties.seq
 import org.scalajs.dom
 
 import scala.concurrent.ExecutionContext
 import scalatags.JsDom.all._
 
 class UdashButtonToolbar[ItemType, ElemType <: Property[ItemType]] private
-                        (val items:io.udash.properties.SeqProperty[ItemType, ElemType],
+                        (val items:seq.SeqProperty[ItemType, ElemType],
                          override val componentId: ComponentId)
                         (itemFactory: (ElemType) => dom.Element) extends UdashBootstrapComponent {
   override lazy val render: dom.Element = {
@@ -54,7 +55,7 @@ object UdashButtonToolbar {
     * @return `UdashButtonToolbar` component, call render to create DOM element representing this toolbar.
     */
   def reactive[ItemType, ElemType <: Property[ItemType]]
-              (items: io.udash.properties.SeqProperty[ItemType, ElemType],
+              (items: seq.SeqProperty[ItemType, ElemType],
                itemFactory: (ElemType) => dom.Element,
                componentId: ComponentId = UdashBootstrap.newId()): UdashButtonToolbar[ItemType, ElemType] =
     new UdashButtonToolbar[ItemType, ElemType](items, componentId)(itemFactory)
