+++ bootstrap/frontend/src/main/scala/io/udash/bootstrap/pagination/UdashPagination.scala
@@ -3,6 +3,7 @@ package pagination
 
 import io.udash._
 import io.udash.bootstrap.UdashBootstrap.ComponentId
+import io.udash.properties.seq
 import org.scalajs.dom
 import org.scalajs.dom.Event
 
@@ -10,7 +11,7 @@ import scala.concurrent.ExecutionContext
 
 sealed trait PaginationComponent[PageType, ElemType <: Property[PageType]] extends UdashBootstrapComponent {
   /** Sequence of pagination elements. Pagination will automatically synchronize with this property changes. */
-  def pages: properties.SeqProperty[PageType, ElemType]
+  def pages: seq.SeqProperty[PageType, ElemType]
 
   /** Index of selected page. */
   def selectedPage: Property[Int]
@@ -30,7 +31,7 @@ sealed trait PaginationComponent[PageType, ElemType <: Property[PageType]] exten
 
 class UdashPagination[PageType, ElemType <: Property[PageType]] private
                      (size: PaginationSize, showArrows: Property[Boolean], highlightActive: Property[Boolean], override val componentId: ComponentId)
-                     (val pages: properties.SeqProperty[PageType, ElemType], val selectedPage: Property[Int])
+                     (val pages: seq.SeqProperty[PageType, ElemType], val selectedPage: Property[Int])
                      (itemFactory: (ElemType, UdashPagination.ButtonType) => dom.Element)(implicit ec: ExecutionContext)
   extends PaginationComponent[PageType, ElemType] {
 
@@ -70,7 +71,7 @@ class UdashPagination[PageType, ElemType <: Property[PageType]] private
 }
 
 class UdashPager[PageType, ElemType <: Property[PageType]] private[pagination](aligned: Boolean, override val componentId: ComponentId)
-                (val pages: properties.SeqProperty[PageType, ElemType], val selectedPage: Property[Int])
+                (val pages: seq.SeqProperty[PageType, ElemType], val selectedPage: Property[Int])
                 (itemFactory: (ElemType, UdashPagination.ButtonType) => dom.Element) extends PaginationComponent[PageType, ElemType] {
 
   override lazy val render: dom.Element = {
@@ -147,7 +148,7 @@ object UdashPagination {
   def apply[PageType, ElemType <: Property[PageType]]
            (size: PaginationSize = PaginationSize.Default, showArrows: Property[Boolean] = Property(true),
             highlightActive: Property[Boolean] = Property(true), componentId: ComponentId = UdashBootstrap.newId())
-           (pages: properties.SeqProperty[PageType, ElemType], selectedPage: Property[Int])
+           (pages: seq.SeqProperty[PageType, ElemType], selectedPage: Property[Int])
            (itemFactory: (ElemType, UdashPagination.ButtonType) => dom.Element)(implicit ec: ExecutionContext): UdashPagination[PageType, ElemType] =
     new UdashPagination(size, showArrows, highlightActive, componentId)(pages, selectedPage)(itemFactory)
 
@@ -165,7 +166,7 @@ object UdashPagination {
     */
   def pager[PageType, ElemType <: Property[PageType]]
            (aligned: Boolean = false, componentId: ComponentId = UdashBootstrap.newId())
-           (pages: properties.SeqProperty[PageType, ElemType], selectedPage: Property[Int])
+           (pages: seq.SeqProperty[PageType, ElemType], selectedPage: Property[Int])
            (itemFactory: (ElemType, UdashPagination.ButtonType) => dom.Element)(implicit ec: ExecutionContext): UdashPager[PageType, ElemType] =
     new UdashPager(aligned, componentId)(pages, selectedPage)(itemFactory)
 }
