+++ bootstrap/frontend/src/main/scala/io/udash/bootstrap/table/UdashTable.scala
@@ -3,13 +3,14 @@ package table
 
 import io.udash._
 import io.udash.bootstrap.UdashBootstrap.ComponentId
+import io.udash.properties.seq
 import org.scalajs.dom
 import org.scalajs.dom._
 
 class UdashTable[ItemType, ElemType <: Property[ItemType]] private
                 (striped: Property[Boolean], bordered: Property[Boolean], hover: Property[Boolean],
                  condensed: Property[Boolean], override val componentId: ComponentId)
-                (val items: properties.SeqProperty[ItemType, ElemType])
+                (val items: seq.SeqProperty[ItemType, ElemType])
                 (headerFactory: Option[() => dom.Element],
                  rowFactory: (ElemType) => dom.Element) extends UdashBootstrapComponent {
 
@@ -57,7 +58,7 @@ object UdashTable {
            (striped: Property[Boolean] = Property(false), bordered: Property[Boolean] = Property(false),
             hover: Property[Boolean] = Property(false), condensed: Property[Boolean] = Property(false),
             componentId: ComponentId = UdashBootstrap.newId())
-           (items: properties.SeqProperty[ItemType, ElemType])
+           (items: seq.SeqProperty[ItemType, ElemType])
            (rowFactory: (ElemType) => dom.Element,
             headerFactory: Option[() => dom.Element] = None): UdashTable[ItemType, ElemType] =
     new UdashTable(striped, bordered, hover, condensed, componentId)(items)(headerFactory, rowFactory)
