+++ bootstrap/frontend/src/main/scala/io/udash/bootstrap/utils/UdashBreadcrumbs.scala
@@ -2,6 +2,7 @@ package io.udash.bootstrap
 package utils
 
 import io.udash.bootstrap.UdashBootstrap.ComponentId
+import io.udash.properties.seq
 import io.udash.{properties, _}
 import org.scalajs.dom
 import org.scalajs.dom.Element
@@ -10,7 +11,7 @@ import org.scalajs.dom.html.Anchor
 import scala.concurrent.ExecutionContext
 
 class UdashBreadcrumbs[ItemType, ElemType <: Property[ItemType]] private
-                      (val pages: properties.SeqProperty[ItemType, ElemType], override val componentId: ComponentId)
+                      (val pages: seq.SeqProperty[ItemType, ElemType], override val componentId: ComponentId)
                       (itemFactory: (ElemType) => dom.Element,
                        isSelected: (ItemType) => Boolean)(implicit ec: ExecutionContext) extends UdashBootstrapComponent {
   override lazy val render: Element = {
@@ -55,7 +56,7 @@ object UdashBreadcrumbs {
     * @return `UdashBreadcrumbs` component, call render to create DOM element.
     */
   def apply[ItemType, ElemType <: Property[ItemType]]
-           (pages: properties.SeqProperty[ItemType, ElemType], componentId: ComponentId = UdashBootstrap.newId())
+           (pages: seq.SeqProperty[ItemType, ElemType], componentId: ComponentId = UdashBootstrap.newId())
            (itemFactory: (ElemType) => dom.Element,
             isSelected: (ItemType) => Boolean = (_: ItemType) => false)(implicit ec: ExecutionContext): UdashBreadcrumbs[ItemType, ElemType] =
     new UdashBreadcrumbs(pages, componentId)(itemFactory, isSelected)
