+++ bootstrap/frontend/src/main/scala/io/udash/bootstrap/utils/UdashListGroup.scala
@@ -3,11 +3,11 @@ package utils
 
 import io.udash._
 import io.udash.bootstrap.UdashBootstrap.ComponentId
-import io.udash.properties.SeqProperty
+import io.udash.properties.seq.SeqProperty
 import org.scalajs.dom
 import org.scalajs.dom.Element
 
-class UdashListGroup[ItemType, ElemType <: Property[ItemType]] private(items: properties.SeqProperty[ItemType, ElemType], override val componentId: ComponentId)
+class UdashListGroup[ItemType, ElemType <: Property[ItemType]] private(items: SeqProperty[ItemType, ElemType], override val componentId: ComponentId)
                                                                       (body: (ElemType) => dom.Element) extends UdashBootstrapComponent {
   import scalatags.JsDom.all._
