+++ bootstrap/frontend/src/test/scala/io/udash/bootstrap/carousel/UdashCarouselTest.scala
@@ -3,7 +3,7 @@ package io.udash.bootstrap.carousel
 import io.udash._
 import io.udash.bootstrap.carousel.UdashCarousel.{CarouselEvent, SlideChangeEvent, SlideChangedEvent}
 import io.udash.bootstrap.{BootstrapStyles, BootstrapTags}
-import io.udash.properties.SeqProperty
+import io.udash.properties.seq.SeqProperty
 import io.udash.testing.AsyncUdashFrontendTest
 import io.udash.wrappers.jquery._
