+++ core/frontend/src/main/scala/io/udash/bindings/CheckButtons.scala
@@ -1,6 +1,7 @@
 package io.udash.bindings
 
-import io.udash.properties.{ReadableProperty, SeqProperty}
+import io.udash.properties.seq.SeqProperty
+import io.udash.properties.single.ReadableProperty
 import org.scalajs.dom.{html, _}
 
 import scala.concurrent.ExecutionContext
