+++ core/frontend/src/main/scala/io/udash/bindings/Checkbox.scala
@@ -1,6 +1,6 @@
 package io.udash.bindings
 
-import io.udash.properties.Property
+import io.udash.properties.single.Property
 import org.scalajs.dom.{html, _}
 
 import scala.concurrent.ExecutionContext
