+++ core/frontend/src/main/scala/io/udash/bindings/Select.scala
@@ -1,6 +1,7 @@
 package io.udash.bindings
 
-import io.udash.properties.{Property, ReadableProperty, SeqProperty}
+import io.udash.properties.single.{Property, ReadableProperty}
+import io.udash.properties.seq.SeqProperty
 import org.scalajs.dom.{html, _}
 
 import scala.concurrent.ExecutionContext
