+++ core/frontend/src/main/scala/io/udash/bindings/modifiers/AttrModifier.scala
@@ -2,6 +2,7 @@ package io.udash.bindings.modifiers
 
 import io.udash.bindings._
 import io.udash.properties._
+import io.udash.properties.single.ReadableProperty
 import org.scalajs.dom
 import org.scalajs.dom._
