+++ core/frontend/src/main/scala/io/udash/bindings/modifiers/SeqAsValueModifier.scala
@@ -1,6 +1,8 @@
 package io.udash.bindings.modifiers
 
 import io.udash.properties._
+import io.udash.properties.seq.ReadableSeqProperty
+import io.udash.properties.single.ReadableProperty
 import io.udash.utils.Registration
 import org.scalajs.dom._
