+++ core/frontend/src/main/scala/io/udash/bindings/modifiers/SeqAsValuePatchingModifier.scala
@@ -1,6 +1,8 @@
 package io.udash.bindings.modifiers
 
 import io.udash.properties._
+import io.udash.properties.seq.{Patch, ReadableSeqProperty}
+import io.udash.properties.single.ReadableProperty
 import org.scalajs.dom
 import org.scalajs.dom._
