+++ core/frontend/src/main/scala/io/udash/bindings/modifiers/SeqPropertyModifier.scala
@@ -2,6 +2,8 @@ package io.udash.bindings.modifiers
 
 import io.udash.bindings._
 import io.udash.properties._
+import io.udash.properties.seq.{Patch, ReadableSeqProperty}
+import io.udash.properties.single.ReadableProperty
 import org.scalajs.dom
 import org.scalajs.dom._
