+++ core/frontend/src/test/scala/io/udash/bindings/CheckButtonsTest.scala
@@ -1,6 +1,6 @@
 package io.udash.bindings
 
-import io.udash.properties.SeqProperty
+import io.udash.properties.seq.SeqProperty
 import io.udash.testing.UdashFrontendTest
 
 class CheckButtonsTest extends UdashFrontendTest {
