+++ core/frontend/src/test/scala/io/udash/bindings/TagsBindingTest.scala
@@ -3,7 +3,7 @@ package io.udash.bindings
 import java.util.concurrent.atomic.AtomicInteger
 
 import io.udash._
-import io.udash.properties.{ImmutableValue, SeqProperty}
+import io.udash.properties.{ImmutableValue, seq}
 import io.udash.testing.UdashFrontendTest
 import org.scalajs.dom.Element
 
@@ -268,7 +268,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
 
   "produce for SeqProperty" should {
     "update content of DOM element" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         span(),
         produce(p)((s: Seq[Int]) => {
@@ -311,7 +311,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "handle null value providing empty Seq to callback" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         produce(p)((s: Seq[Int]) => {
           div(s.map(v => {
@@ -358,8 +358,8 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "not swap position" in {
-      val p = SeqProperty[Int](1, 2, 3)
-      val p2 = SeqProperty[Int](3, 2, 1)
+      val p = seq.SeqProperty[Int](1, 2, 3)
+      val p2 = seq.SeqProperty[Int](3, 2, 1)
       val template = div(
         "A",
         produce(p)((s: Seq[Int]) => {
@@ -394,7 +394,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "work after moving element in DOM" in {
-      val p = SeqProperty[String]("A")
+      val p = seq.SeqProperty[String]("A")
       val b = span(produce(p)((v: Seq[String]) => span(v.mkString).render)).render
       val template = div(b).render
       val template2 = emptyComponent()
@@ -424,7 +424,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
 
   "Patching produce for SeqProperty" should {
     "init and update content of DOM element" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         span(),
         produce(p,
@@ -472,7 +472,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "handle null value providing empty Seq to callback" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         produce(p,
           (seq: Seq[Property[Int]]) => div(seq.map(p => span(s"${p.get} ")): _*).render,
@@ -523,7 +523,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "not swap position" in {
-      def prod(p: SeqProperty[Int, Property[Int]]) = {
+      def prod(p: seq.SeqProperty[Int, Property[Int]]) = {
         produce(p,
           (seq: Seq[Property[Int]]) => div(seq.map(p => span(p.get)): _*).render,
           (patch: Patch[Property[Int]], elem: Element) => {
@@ -536,8 +536,8 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
         )
       }
 
-      val p = SeqProperty[Int](1, 2, 3)
-      val p2 = SeqProperty[Int](3, 2, 1)
+      val p = seq.SeqProperty[Int](1, 2, 3)
+      val p2 = seq.SeqProperty[Int](3, 2, 1)
       val template = div(
         "A",
         prod(p),
@@ -564,7 +564,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
 
   "repeat" should {
     "update content of DOM element" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         span(),
         repeat(p)((p: Property[Int]) => {
@@ -685,7 +685,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "handle null value providing empty text placeholder" in {
-      val p = SeqProperty[Int](1, 2, 3)
+      val p = seq.SeqProperty[Int](1, 2, 3)
       val template = div(
         repeat(p)((p: Property[Int]) => {
           val v = p.get
@@ -774,7 +774,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "not swap position" in {
-      def rep(p: SeqProperty[Int, Property[Int]]) = {
+      def rep(p: seq.SeqProperty[Int, Property[Int]]) = {
         repeat(p)((p: Property[Int]) => {
           val v = p.get
           if (v % 2 == 0) b(v.toString).render
@@ -782,8 +782,8 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
         })
       }
 
-      val p = SeqProperty[Int](1, 2, 3)
-      val p2 = SeqProperty[Int](3, 2, 1)
+      val p = seq.SeqProperty[Int](1, 2, 3)
+      val p2 = seq.SeqProperty[Int](3, 2, 1)
       val template = div(
         "A",
         rep(p),
@@ -838,7 +838,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "not swap position with CallbackSequencer" in {
-      def rep(p: SeqProperty[Int, Property[Int]]) = {
+      def rep(p: seq.SeqProperty[Int, Property[Int]]) = {
         repeat(p)((p: Property[Int]) => {
           val v = p.get
           if (v % 2 == 0) b(v.toString).render
@@ -846,8 +846,8 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
         })
       }
 
-      val p = SeqProperty[Int](1, 2, 3)
-      val p2 = SeqProperty[Int](3, 2, 1)
+      val p = seq.SeqProperty[Int](1, 2, 3)
+      val p2 = seq.SeqProperty[Int](3, 2, 1)
       val template = div(
         "A",
         rep(p),
@@ -883,7 +883,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
       case object EvensFilter      extends NumbersFilter(i => i % 2 == 0)
 
       val filter = Property[NumbersFilter]
-      val numbers = SeqProperty[Int]
+      val numbers = seq.SeqProperty[Int]
 
       filter.set(OddsFilter)
       numbers.set(Seq(1, 2, 3, 4, 5))
@@ -922,7 +922,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "work with filtered transformed SeqProperty" in {
-      val doubles = SeqProperty[Double](1.5, 2.3, 3.7)
+      val doubles = seq.SeqProperty[Double](1.5, 2.3, 3.7)
       val ints = doubles.transform((d: Double) => d.toInt, (i: Int) => i.toDouble)
       val evens = ints.filter(_ % 2 == 0)
 
@@ -962,7 +962,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
       case object EvensFilter      extends NumbersFilter(i => i % 2 == 0)
 
       val filter = Property[NumbersFilter]
-      val numbers = SeqProperty[Int]
+      val numbers = seq.SeqProperty[Int]
 
       filter.set(OddsFilter)
       numbers.set(Seq(1, 2, 3, 4, 5))
@@ -1016,7 +1016,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
                       override val completed: Boolean) extends TodoElement
 
       val filter = Property[TodosFilter]
-      val todos = SeqProperty[TodoElement]
+      val todos = seq.SeqProperty[TodoElement]
 
       val done = todos.filter(CompletedTodosFilter.matcher)
       val patches = scala.collection.mutable.ArrayBuffer.empty[Patch[_]]
@@ -1110,7 +1110,7 @@ class TagsBindingTest extends UdashFrontendTest with Bindings { bindings: Bindin
     }
 
     "work after moving element in DOM" in {
-      val p = SeqProperty[String]("A")
+      val p = seq.SeqProperty[String]("A")
       val b = span(repeat(p)((v: Property[String]) => span(v.get).render)).render
       val template = div(b).render
       val template2 = emptyComponent()
