+++ core/frontend/src/test/scala/io/udash/properties/PropertyTest.scala
@@ -1,5 +1,8 @@
 package io.udash.properties
 
+import io.udash.properties.model.ModelProperty
+import io.udash.properties.seq.{Patch, ReadableSeqProperty, SeqProperty}
+import io.udash.properties.single.{Property, ReadableProperty}
 import io.udash.testing.UdashFrontendTest
 
 import scala.collection.mutable
