+++ core/frontend/src/test/scala/io/udash/properties/PropertyUsageTest.scala
@@ -3,6 +3,11 @@ package io.udash.properties
 import io.udash.testing.UdashFrontendTest
 
 class PropertyUsageTest extends UdashFrontendTest {
+  // DO NOT REMOVE THIS IMPORTS!
+  import io.udash.properties.single._
+  import io.udash.properties.model._
+  import io.udash.properties.seq._
+
   "CastableProperty" should {
     "allow safe casting to ModelProperty or SeqProperty " in {
       """val p = Property[Int]
