+++ core/macros/src/main/scala/io/udash/macros/PropertyMacros.scala
@@ -15,14 +15,14 @@ class PropertyMacros(val c: blackbox.Context) {
   val PropertyCreatorCls = tq"$Package.PropertyCreator"
   val PropertyCreatorCompanion = q"$Package.PropertyCreator"
 
-  val ReadablePropertyCls = tq"$Package.ReadableProperty"
-  val PropertyCls = tq"$Package.Property"
-  val DirectPropertyImplCls = tq"$Package.DirectPropertyImpl"
-  val SeqPropertyCls = tq"$Package.SeqProperty"
-  val DirectSeqPropertyImplCls = tq"$Package.DirectSeqPropertyImpl"
-  val ModelPropertyCls = tq"$Package.ModelProperty"
-  val ModelPropertyImplCls = tq"$Package.ModelPropertyImpl"
-  val CastablePropertyCls = tq"$Package.CastableProperty"
+  val ReadablePropertyCls = tq"$Package.single.ReadableProperty"
+  val PropertyCls = tq"$Package.single.Property"
+  val DirectPropertyImplCls = tq"$Package.single.DirectPropertyImpl"
+  val SeqPropertyCls = tq"$Package.seq.SeqProperty"
+  val DirectSeqPropertyImplCls = tq"$Package.seq.DirectSeqPropertyImpl"
+  val ModelPropertyCls = tq"$Package.model.ModelProperty"
+  val ModelPropertyImplCls = tq"$Package.model.ModelPropertyImpl"
+  val CastablePropertyCls = tq"$Package.single.CastableProperty"
   val CallbackSequencerCls = q"$Package.CallbackSequencer"
 
   val PatchCls = tq"$Package.SeqProperty.Patch"
